find: fix sorting
Bareos internal logic relies on directories always coming _after_ all
files they include.  As such we should not reorder files and
directories at all since its hard to guarantee that one is not
includede in the other.

The idea is to first partition regular files and directories into two
distinct piles (in a stable way so that the directory structure does
not get damaged).  Afterwards we can safely sort the files to
hopefully speed up reading them.