bsr: make conditional easier to understand
Testing whether two intervals intersect -- which is what the
conditional is trying to do -- can be easily done with

a <= d && b >= c,

where the intervals are [a, b] and [c, d] with a<=b and c<=d.
It is then easy to check that this is equivalent to

max(a, c) <= min(b, d),

which is what the code now does.