append: introduce MessageReceiver
This class reads data from the given socket asynchroniously and makes
them available in a preparsed manner on the main thread.  It is
important that the given socket is only used by one thread at a time,
as such, if you create a MessageHandler for a certain socket, you need
to take care to not use that socket at all while the message handler
is alive.