build: remove jansson dependency
Previously, Bareos was built for MacOS with Jansson support. As we did
not ship the Jansson library, this lead to mostly unusable binaries.
This change disables Jansson when building for MacOS and also creates a
list of dependencies of all the binaries so we can review them after the
build.