restore-multi: fix hardlink related problems on jenkins
Selecting files with dir= (and file=) does not autoselect original
files in case a hard linked file is marked.   This can lead to
problems in this test if the first hard linked file that is
encountered is not one inside our to-be-restored directories.

This commit fixes this by instead restoring very simple non-empty
directories (because this also makes a difference to the directory=
command) that do not contain any hard links or soft links.