vadp-dumper: add support for VDL_QueryAllocatedBlocks()
This function returns -- up to a certain amount of precision -- a list
of intervals of all sectors containing useful (i.e. allocated) data.

For full backups its recommended to backup only the allocated blocks;
for other kinds of backups (incremental/differential) it is only
necessary to backup those sectors that are both changed & allocated.

Since we currently cannot take advantage of information regarding
unallocated, changed blocks, we just ignore them.  In the future these
could be used for faster restores & consolidation.

The algorithm used for finding out the intersection is trivial as both
lists are sorted (and the intervals themselves are disjoint in each
set).  As such it is enough to just go through both arrays linearly at
the same time, sending the pairwise intersection if any, and finally
advancing  the pointer to the "smaller" interval.