xxhash: automatically choose best way to build
Previously, we blindly used xxh3.h to have xxhash inlined completely.
While this provides decent performance, we can do better.
Based on wheter using bundled xxhash and the target platform, we
automatically select which header to include to get either the
dynamically linked system-provided library, an optimized algorithm
based on the CPU's features selected at runtime or just the plain
variant of xxhash.

We also configure the bundled xxhash library to be built with -O3.