python-fd: make failed plugin-load fatal
Just returning bRC_Error from load_bareos_plugin() is apparently not
enough to make Bareos fail the backup. Thus we emit an M_FATAL instead
of an M_ERROR to fast-fail the job.
This also fixes a subtle bug in the class decorator: as we weren't
returning the class, the class itself was not available under its name
which could lead to strange effects.