copy-thread: add capacity check
Since the input size is a variable, we need to ensure that we hold
enough capacity to actually read the data into the buffer.  Otherwise,
if we get unlucky, we will run into the situation were the first few
cbts are very small, and the rest are really big, ensuring a buffer
overflow; possibly causing a crash.