checkpoints-test: fix comparing bad numbers
The test was comparing the number of files that the fd send to the
sd (this is the number displayed in the backup summary) to the number
of files that get restored (which is equal to the number of files that
the sd sends to the director).  This of course does not make sense and
can cause problems if the cancel does not happen fast enough.  Since
in that case the fd might continue sending files to the sd with the sd
never having the chance to send those to the director.

The fix is simple: we need to check the backup log and add up all
checkpoint batch inserts instead.  This should be equal to the number
of files that the sd send to the director.

For easier debugging I also added tracing to the test.