remove explicit dependency on pgsql
USES+=pgsql will pin us to a specific version of PostgreSQL, which is
probably not a great idea, so we remove it.
This will retain the (automatic) requirement to libpq, so a compatible
postgresql client will be required by the binary package.

(cherry picked from commit 91f1842683f22abdf0cd7be11690dea24da51e77)