fvec: remove fallocate
It turns out that posix_fallocate can become very slow on glibc if you
the filesystem does not really support it since the fallback is just
writing a bunch of zeroes into the file.   This is not what we want so
we only use ftruncate for now.

(cherry picked from commit 5436ec996a957a41ef9b8cf70bc840199a99badb)