pr_tool: make find_user_remote() safer
previously we only considered the first URL for a remote. With this
change we ignore remotes that have multiple URLs configured and emit a
warning.

Remotes with multiple URLs are only used if you want to push to multiple
locations at once, which is probably not what `pr_tool backport` should
be doing.