cmake: add DebugEdit module
by using the DebugEdit module we add a POST_BUILD command to each CMake
target that generates ELF binaries if we have debugedit available.
This command will basically undo the job of -fdebug-prefix-map.
The idea here is that we want to have all object files with canonical
paths in ccache, but the resulting binaries should have the
non-canonical path embedded so debuginfo/debugsource generation will
work correctly.