reserve: switch to std::mutex
reservations_lock_count was not used (and was used improperly), so it
was removed.  reservation_lock is defined as rwlock, but only ever
used as a normal lock, so it was replaced by a std::mutex.

(cherry picked from commit 5bf20603152ea122e2ff6f0d753f45d12ca8f79a)