Fix warning about missing delcandidates table in director.
Pruning jobs would cause the following warning in the output of the
director (e.g. visible through
“systemctl status bareos-director.service”):

  table "delcandidates" does not exist, skipping

This warning is harmless but clutters the output. With this commit, this
problem is fixed by disabling warnings for the PostgreSQL connection
that is established by the director.

See https://bugs.bareos.org/view.php?id=1561 and
https://bugs.bareos.org/view.php?id=1606.