VMware Plugin: Backup and Restore NVRAM
The NVRAM of VMs is now backed up and restored when a VM is recreated.
This ensures that EFI enabled VMs will boot successfully.

Additionally, the exception handling and error messages in the functions
connect_vmware() and retrieve_vcthumbprint() were enhanced.

The plugin now makes sure to store all files belonging to a recreated VM
in the same directory in VMFS.

This commit also contains some changes to improve the pylint score.