percona-xtrabackup: fix bug with subprocess call
Fix non-terminating subprocess behaviour in case of bareos-storage
unavailable during a Backup job.

This commit terminates the subprocess in case bareos-storage is
unavailable for some reason during a Backup job.
This is a bug fix because the previous behaviour causes high memory
usage that may cause OOM in some cases.

Move the logic of terminating the xtrabackup(dumper) subprocess
in case of IO_CLOSE to separate function end_dumper().
Add timeouts and change the kill signal to terminate.
Add convenient debug and job messages.