VMware Plugin: Adapt to older urllib3 versions
In urllib3 version 1.26 the parameter method_whitelist was replaced by
allowed_methods. The Plugin now detects which parameter to use so that
is also works with older versions.

Also the new plugin option fallback_to_full_cbt was added, which can be
used when it is desired to run a new full level job anyway when
incremental CBT is not possible.