stored: fix not respecting bsr volume order
The way it worked was the following:
1) Go through all bsr, and note which volumes are used (basically in a
set)
2) For each listed volume: open it, position it to the first position
mentioned in any bsr entry, and read from it until _no_ bsr entry
wants to read from it anymore
3) send all records that match _any_ bsr entry to the callback

The Problem with this approach is that we do not guarantee that the
order of records send to to e.g. the fd stays the same.  In some
situations we could even interleave records from two different files!
Think of a volume separated into blocks like so:

| job 1 | job 2 | job 1 | job 2 | job 1 | job 2 |

where job 1 and job 2 are jobs that backed up to the
same device at the same time.  If you now try to restore both of them,
the records will arrive interleaved!  Cleary we should first send all
of job 1 blocks and only then start sending all of job 2 blocks.

This is fixed in the following way:

1) We know create an ordered list of volumes by either appending the
bsrs volume to the end of the list or merging it into the last entry of
the list in case they refer to the same volume. E.g. the bsrs

| Vol 1 | Vol 2 | Vol 1 | Vol 1 |

will result in the list Vol 1 -> Vol 2 -> Vol 1.

2) We go through the list of volumes, keeping track of the current bsr
entry.  We only accept records of the _current_ bsr entry.  If we
finished the current bsr entry, and the next bsr entry uses the same
volume, we set the current bsr to the next bsr (This causes the core
to reposition the current volume to the start of that bsr).  Otherwise
we mount the next volume and then switch bsrs.

3) We only send records to the callback that match the current bsr.

This should hopefully prevent any interleaving.