python-fd: improve BareosFdWrapper
The wrappers won't catch any exceptions, but will let python-fd's
error handler take care of them. Previously this was a bit unclear as
load_bareos_plugin() did handle exceptions, but the other wrapper
functions did not.

We now declare what names will be exported to consuming plugins. This
ensures we can change the implementation as long as we don't break the
API. This also ensures that we don't leak implementation details into
the plugins using the wrapper.

improve compliance with pylint. Some warnings can only be ignored.
Especially naming issues in the API are not fixable in a
backwards-compatible manner.