alist: fix use of internal iterator
As alists mutated themselves when iterated over, it was not safe to
even iterate over them from multiple threads at the same time.  This
lead to hard to debug bugs.  This commit changes the foreach_alist
macro in such a way that it does not rely on the internal iterator
state any longer and instead uses a simple pointer pair.