devtools: accept "bug" label as alias of "bugfix"
We currently have "bugfix" and "bug" as labels in the Bareos project.
The "bugfix" label was added for PRs that fix a bug. However, since we
started using issues we also have the "bug" label for issues that
descibe a bug. We should probably use "bug" for everything, but for
that pr-tool needs to accept "bug" as an alias for "bugfix" when
generating changelog records.