systemtests: replace config-syntax-crash test
Invalid chars in WhereAcls are now properly marked as configuration
errors. Therefore in the original test config-syntax-crash, the Bareos
Director refuses to start (with a proper error message), instead of just
ignoring the invalid entry.
Instead of adapting the config-syntax-crash, the test for invalid chars
is now added to the python-bareos tests and the config-syntax-crash
removed.