python-fd: serialize calls to PyThreadState_New with GIL
I am not sure if this is necessary, but the documentation says:

```
The global interpreter lock need not be held, but may be held if
it is necessary to serialize calls to this function.
```

But it does not elaborate on *when* it is necessary to serialize those
calls.  As such I will pretend that it is also necessary and use the
gil to do so.

(cherry picked from commit d8fa574a624e15cb424836d99c590582137b0057)