python-fd: fix calling bareosfd functions to early
While it should technically be no problem to call the bareosfd capi
without holding the gil, we have to keep in mind that we are still
modifying shared global state and as such we need to serialize those
calls.  (Ab-)Using the gil to do so is the easiest way to fix these
problems.

I am not 100% sure how running multiple plugin jobs in parallel never
ran into issues considering that handlePluginEvent was constantly
clobbering the global "plugin context" in bareosfd.  Its possible that
im overlooking something.