db: fix not locking the db to print its error
db->strerror() should only be called with the lock held unless the db
is private.  90% of the time db calls look like
```
if (!db->Function()) {
   Error(db->strerror());
}
```
In these cases we can need to lock the db during the if, otherwise we
could read stale info (or even run into a race condition as strerror
is unsynchronized).