findlib: refactor xattr_build_data_t
instead of keeping a ptr to `xattr_build_data_t` in `XattrData`, we now
have to distinct types `XattrData` and `XattrBuildData` with the latter
inheriting from the former and providing all members of
xattr_build_data_t (and proper ctor/dtor).

The downside is that some of the functions in xattr.cc that were passed
a `XattrData*` before will now need a `XattrBuildData*` so we need to
downcast when calling these.