chunked_device: fix bug with chunk data loss
Enqueing the copy of a chunk while the original chunk was not yet
flushed, is a no-op. When inserting the new io request, only volume name
and chunk number were compared. Then UpdateChunkIoRequest() was called
which only updates the existing io request if the buffers are the same -
which they're not, as the new chunk contains a copy of the original
buffer.

We now properly replace the existing io request with the new one, which
will only upload the correct chunk data.