chunked_device: fix bogus selection of next chunk
When reopening a volume where the first chunk has data that was not yet
flushed, the device will setup a new chunk with a copy of the buffered
chunk data.
However, in this case the end_offset is bogus, which would in turn make
subsequent write requests think the chunk is full and make them skip to
the next chunk.
This fix sets the correct end_offset so the chunk is not considered
full.