VMware Plugin: Suggested changes for PR#1876
add_disk_devices_to_vm() now always sets the backing filename to
contain the datastore name only, so that won't the exception
vim.fault.FileAlreadyExists won't happen. Instead any unexpected
exception when adding a disk is handled now with a limited number
of retries. Also a job message is emitted, when the backing
filename of the created disk does not match the expected one.

Renamed the plugin option to restore_allow_disks_mismatch, but with
the change described above, it's probably not needed even when
using VSAN.