lib: child-btimers: add dead-man-switch
You can now signal a child-timer that the child is behaving correctly,
even though it won't finish within the timeout.
By calling TimerChildOperatesProperly(btimer_t&) a flag will be set that
will make the timer defer the killing of the child.

This commit also adds some basic testing for bpipe.