stored: dplcompat: don't timeout while doing I/O
Previously, when upload or download operations did not finish within the
hard-coded timeout of 30 seconds, the process was killed and the
operation considered a failure.
We now reset the timeout after every read and write operation. So as
long as some I/O is happening at least every 30 seconds, the operation
will not time out.