chunked_device: flush all volumes in is_written()
when ReleaseDevice() calls d_flush() on the volume, it should only
return if all buffers were flushed correctly. However, as a volume
change does not imply d_flush(), we might have chunks from a
previously mounted volume still in the queue.
This change ensures that not only the chunks of the current volume are
flushed, but that the queue is completely empty.