systemtests: disable mariabackup test on SUSE
We currently disable the test on suse as the plugin does not support
loading a local defaults file when the MySQLdb module is found. As
the test only works with that defaults file (and we do not want to
potentially overwrite the current users configuration), we instead
disable the test.

Signed-off-by: Bruno Friedmann <bruno.friedmann@bareos.com>
(cherry picked from commit 5558563e8f3453c44e943fa9ade295e4dfc1d197)