systemtest python-bareos: speed up tests by 25%
The python-bareos systemtests do test a lot of functionality.
As every test must be able to run indepently and in random order,
the tests often ensured that test data is available
by running a backup job.

This change uses the recently introduced function get_backup_jobid
to ensure the availability of test data in a more efficient way:
if a job with the requested parameter already exists,
it's jobid is returned. Otherwise the requested job is started.