systemtests: fix ordering and dependencies
in ctest there is no way to permanently hold a resource lock while a
fixture is set up.
Thus there could be cases where the different autochanger tests, even
though there is a lock in place, would run intermixed which doesn't
work. This now adds simple dependencies to make sure the next test using
an autochanger waits for the cleanup of another one.