stored: replace PRE_LABEL with VOL_LABEL
instead of writing a pre-label to an empty volume when labeling it, we
write the real volume-label right away so we don't have to rewrite it
when we actually start writing to the volume.
This will allow using WORM storages.

(cherry picked from commit ee8831b1635ec60d26e53d7b35d7a4d58c967e72)