stored: change "previously written" message in mount.cc
When mounting a volume that contains a VOL_LABEL, the SD would log the
following message:
Volume "XXX" previously written, moving to end of data.

As that could be misleading now that newly labeled volumes will contain
a VOL_LABEL and the message would appear for volumes that only habe a
label, but no data.