tests: fix reservation tests by disabling jit reservation
The reservation related tests cannot run with jit reservation as they
try to check for an actual reservation which may not have taken place
yet; or was announced differently.

Either way the test will fail with jit reservation = Yes, so we need
to explicitly disable it as it is the default now.

(cherry picked from commit c4522636e04d5b34ab3bd56e7ffdeb2c061ff848)