mount: add quickfix for autolabel issue
We currently overwrite the current volume info on the device with the
outdated volume info from the director (which has size = 0) before
"updating" the volume info on the director.  This causes the volume
size mismatch now that we do not have the pre_label split
anymore (previously we would execute RewriteVolumeLabel before writing
to the device which would have fixed this issue).

As a quick fix we can just disable this overwriting and just send the
real info to the director immediately.

(cherry picked from commit 93077a2add83a9b0e39d4b87a42e56ff7aee1b0d)