python-fd: fix version hex
Our python version hex was not correct for numbers >= 10.  We
previously emitted e.g. 0x03100000 for 3.10, when its supposed to be
0x030A0000.

This was causing weird crashes because workarounds for old versions
that were not supposed to be enabled for newer versions were in fact
enabled.

(cherry picked from commit 9fc3ae5d922a48b1c91ffe561dcd2085f1521b07)