restore: remove double cancel
NativeRestoreCleanup is already done on error in job_thread (job.cc)
in case an error occured, so there is no need to do it twice.

This may cause a crash because pthread_cancel and C++ dont mix well,
especially if we cancel the same thread in a short time interval.
(pthread_cancel causes an exception + destructor calling, which may
lead to an exception being caused inside a destructor, leading to a
crash.)