ua-output: fix resource order
Previously the iteration order depended on the c++ lib
implementation (and could also be change from run to run).  This lead
to issues where 'show job' would print jobs on linux but jobdefs on
windows.

To make sure that always the same element is chosen, we need to make
sure the order is always the same.  This is easily done by just using
a vector here.