btape: adapt to no-PRE_LABEL change
move to 1:0 instead of 0:1 when doing multi-tape unfill.
As we now have a file-mark right after the label block, btape needs to
move to the first block after that file-mark instead of the second block
in the first file.
This change also adds reading the volume label of the second tape, which
was not done before.