chunked_device: fix bug in ChunkedVolumeSize()
While writing to a chunk, a call to ChunkedVolumeSize() would not return
the correct size, as it got the size by either looking at the upload
queue or asking the backend storage.
The patch adds a third option: if the current chunk has been written,
its last offset represents the end of the volume.