grpc-python-module: remove global grpc connection
We need to make sure that the server never gets shutdown on a worker
thread, otherwise the application will deadlock.  As such we cannot
have the server be a global variable, otherwise a call to exit() on a
worker will cause that destructor to run on that thread, leading to
the aforementioned deadlock.