[fixup] make tests work
The previous approach does not work, as the actual path is used during
configuring, meaning that we setup the normal test twice but never the
grpc test.

To fix this we now use the following schema:
If you add a python systemtest, then we not only configure that
systemtest, but also a separate grpc one.  This way we actually have
two different tests with two different names (py3plug vs py3grpc)!