parallel-test: fix reservation order check
On slow/high usage systems, this test may fail because bareos takes to
long to actually start the regular job.  As such we cannot use the
joblog to check if the regular job needed to wait or not, as the
sleepy job mightve already started when we start checking.

Instead we now consult the database to find out if the regular job
needed to wait at all.