systemtests: fix locale settings
By default (C) programs start with the 'C' locale.  This can cause
python to think that the locale is actually an ascii one and will
refuse to work with unicode paths.

We know set the locale to an actual UTF-8 one, so that python will
behave.  This allows us to reenable previously broken tests.