debian: Add missing build dependencies
The new droplet compatible storage backend introduced in #1790 adds two
new dependencies, that were not added to the debian control files. This
commit adds the appropriate packages as build dependencies. Without this
cmake will try to load the dependencies by itself and will fail, because
git is not available by default during the package build.

Please note: If you are building for bullseye, you will either need a
backport of libcli11-dev and libmsgsl-dev or you will need to have git
available during the packaging.