fd-plugins: enable auto grpc
When a plugin cannot be found, the grpc plugin is loaded and a
grpc compat module is defined, then we try to load that plugin with
that compat module.

To facilitate this we must make sure that the plugin handling is able
to modify the plugin command strings both into the plugin and out of
the plugin as we need to add/remove the grpc prefix.

We also remove the `reverse` option of `GeneratePluginEvent` as it was
unused and would mess with the (already complicated) grpc fallback
logic.

(cherry picked from commit 047366373d5f68f5cae17e8851313fd7b7667256)