append: fix accessing null pointer
Currently, if jit reservation is used but no device could be found, we
break out of the backup loop.  Afterwards the sd tries to destroy the
dcr if it exists, which includes releasing the acquired device.  This
is done even if no device was attached to the dcr.

(cherry picked from commit d506a49cb509969f564ec46b6f9618eeee85909d)