stored: remove code paths that ignore reservation
All jobs reserve a viable volume to use when they reserve a device.
Sadly some code paths ignored this and instead just used the currently
mounted device for writing regardless of whether it was reserved by
somebody else or not.

This caused weird `Volume X wanted by Device Y is in use by device Z`
error messages as device Y actually reserved volume X, but device Z
simply decided to use it as it was currently mounted, which ultimately
cause an operator message to be sent out.

As volumes that are currently mounted in the device are preferred for
reservation anyways, we simply remove the `IsSuitableVolumeMounted()`
code paths.

If this becomes a problem in the future we can simply try to reserve
the currently mounted volume and use it if it worked.

(cherry picked from commit 7bb4eb95a169a75796740d04a9c80780c00871b7)