dedup: Fix missmatching types of config and fvec
This fixes some conversion warnings when compiling bareos for 32bit.
The types of config::block_file::End and config::part_file::End are
not the same as fvec<block>::size_type or fvec<part>::size_type for
32bit, beacuse size_t is not uint64_t.

This change does not change the types of the structures, because these
values are serialized on the volume.