VMware Plugin: Requested changes for PR#2152
The UUID part of the changeId from the restoreobject is now compared
with the UUID part of the changeId from the current snapshot. A mismatch
indicates that a new full level job run is required, so in that case the
job will fail. The optional script reschedule_job_as_full.sh can be used
to trigger new full jobs in that case automatically, as that is
currently not possible from the plugin itself.

(cherry picked from commit 617129e90a2b745d0f00605c1bc013f8bff179bf)