systemtests: add second chance for zombie processes
Our systemtests will quit when `waitpid` sees a zombie process.  This
can happen when the system is under a lot of load.  Because of this we
wait for 1 second after discovering a zombie process to give the
system some time to clean it up.

(cherry picked from commit 4e69dd3bfe77c1931e4ea34ee7f934dd991b0fbb)