cmake: better MSVC warning flags
tune warnings flags for MSVC to be a bit more in line with what we use
on GCC/Clang. Especially, ignoring conversion warnings.
This should leave us with a set of warnings that we shoud actually work
on.

(cherry picked from commit 7ba43c48f8207021a3423602af5f4fc0368d8998)