ctest: smart parallelization for tests
as this cannot be determined ahead-of-time for every build/platform it
is probably better to rely on CTEST_PARALLEL_LEVEL being set in the
environment.
Thus, instead of statically setting this to 10, we force 5 for
sanitizer builds and otherwise set it to 10 unless the user has set
CTEST_PARALLEL_LEVEL in the environment.