VMware Plugin: Handling of empty changeId
Under rare circumstances it can happen that the changeId in the backing
information of a snapshotted virtual disk is empty. This should normally
not happen, as it makes subsequent incrementals impossible. Such an
empty changeId is now handled appropriately so that a warning is emitted
if it is empty and when incrementals are attempted the error message was
adapted so that the helper script from
contrib/misc/reschedule_job_as_full
can be used to automatically trigger a full level job in that case.