cmake: disable lto when linking gtests
this disables link-time-optimization when linking our unittests. This is
only possible if we have fat lto objects, so we only do it then.
The benefit of this is that we save a lot of compile-time on programs
that will only run a few times during testing and then will be
discarded.