core: fix not keeping alive configuration
When we copy configuration pointers from one jcr to the next, we need
to make sure that the configuration actually stays alive.

To do this we simply copy the configuration container from the old jcr
into the new jcr, as that container keeps alive everything that we
need.

Maybe we should instead try to match the resources to their new
counterpart ?