grpc: fix leaks
Make sure to reuse/free the object/name memory.

Also enhance the restoreobject test.  If the restoreobject was messed
up, then python may throw an exception when we try to print
it (i.e. when the object has internal zero bytes)!

This is fixed by just printing everything in hex instead.