grpc-impl: fix always adding an additional /
Its not well specified wether _both_ link _and_ fname should end in /,
or if just link should end in /.  The core mostly does it in a way
where fname does _not_ end in /, but link does.

I assumed that this was the rule, but it turns out that the plugins
simply set fname _and_ link to the same path ending in /.

This mismatch of assumption and reality meant that the grpc plugin now
created paths with double / at the end, which tripped up the accurate
information as well as the file listings.

This commit fixes this by always stripping any trailing / from fname
and always adding one trailing / to link (after stripping all the rest
of them).

(cherry picked from commit df97e8e3619f1e6546acdaa51780bdf816256c87)