freebsd: try to keep @sample file during upgrade
based on FreeBSD man page pkg-script the script order on upgrade is:
     1.   new pre-install
     2.   old pre-deinstall
     3.   replace binaries
     4.   new post-install

Based on this, we implemented saving the old configuration
(@sample files that are no longer defined) before they got deleted by
post-deinstall.

Spoiler:
this does not work, as on upgrade the old package gots complete removed,
including running post-deinstall, before any script of the new package
is called.