cmake: enable optimisations on windows
If you execute `RemoveNDebugFlag` before `project()` has the chance to
setup the flags, then `project()` will not set up the flags at all and
they will be empty.

On our non-windows builders that wasnt a problem since we do set the
flags explicitly there, but since we do not do so on our windows
builders, this meant that they were not compiled with any optimisation
flags at all.